refactor: absorb env, hooks, git, and preflight into session (#114)
Removed four satellite modules (env, hooks, git, preflight) from lib.rs
and inlined their logic as private functions in src/run.rs. Each module
was too thin to justify its own file — all code paths converged in
RunSession::prepare(). No public interface changes; prepare() and
execute() signatures are unchanged.

Integration tests for the four modules moved to inline unit tests under
run::tests, preserving the same behavioral coverage.
