+++ src/run.rs
@@ -1,23 +1,20 @@
-use anyhow::{Context, Result};
-use capsule::config::{resolve, CliOverrides, Config, GithubScope, ResolveMode};
+use anyhow::{bail, Context, Result};
+use capsule::config::{resolve, CliOverrides, Config, GitIdentity, GithubScope, ResolveMode};
 use capsule::container_execution::{
     container_name_for, detect_compose_network, post_stream_error, run_container, run_iteration,
     token_remaining_minutes, ExecutionConfig, IterationOutcome,
 };
-use capsule::env::{load_dotenv, parse_dotenv, resolve_gh_token};
-use capsule::git::resolve_git_identity;
-use capsule::hooks::run_before_all;
 use capsule::image_build::{build_base_image, build_derived_image, BuildConfig};
 use capsule::pipeline::{
     build_summary_artifact, PipelineExecutor, PipelineState, RunSummary, StageRunner,
     TerminalReason,
 };
-use capsule::preflight::{check_docker, env_gitignore_warning};
 use capsule::update_check;
 use capsule::verdict::Verdict;
 use std::collections::HashMap;
 use std::io::Write;
 use std::path::{Path, PathBuf};
+use std::process::{Command, Stdio};
 use std::sync::{Arc, Mutex};
 use std::time::SystemTime;
 
@@ -626,6 +623,192 @@ fn resume_hint(session_id: Option<&str>, reason: &TerminalReason) -> Option<Stri
     }
 }
 
+// ── env ──────────────────────────────────────────────────────────────────────
+
+fn parse_dotenv(content: &str) -> HashMap<String, String> {
+    let mut map = HashMap::new();
+    for line in content.lines() {
+        let line = line.trim();
+        if line.is_empty() || line.starts_with('#') {
+            continue;
+        }
+        if let Some(eq_pos) = line.find('=') {
+            let key = line[..eq_pos].trim().to_string();
+            let raw_val = line[eq_pos + 1..].trim();
+            let value = strip_quotes(raw_val).to_string();
+            if !key.is_empty() {
+                map.insert(key, value);
+            }
+        }
+    }
+    map
+}
+
+fn strip_quotes(s: &str) -> &str {
+    if (s.starts_with('"') && s.ends_with('"')) || (s.starts_with('\'') && s.ends_with('\'')) {
+        &s[1..s.len() - 1]
+    } else {
+        s
+    }
+}
+
+fn load_dotenv(capsule_dir: &Path) -> Result<()> {
+    let path = capsule_dir.join(".env");
+    if !path.exists() {
+        return Ok(());
+    }
+    let content =
+        std::fs::read_to_string(&path).with_context(|| format!("reading {}", path.display()))?;
+    for (key, value) in parse_dotenv(&content) {
+        if std::env::var(&key).is_err() {
+            // SAFETY: called before any threads are spawned (ctrlc handler registered later in main).
+            unsafe { std::env::set_var(&key, &value) };
+        }
+    }
+    Ok(())
+}
+
+fn resolve_gh_token(
+    scope: &GithubScope,
+    pre_dotenv_env: &HashMap<String, String>,
+    dotenv_map: &HashMap<String, String>,
+) -> Result<String> {
+    match scope {
+        GithubScope::Local => dotenv_map
+            .get("GH_TOKEN")
+            .filter(|t| !t.is_empty())
+            .cloned()
+            .ok_or_else(|| {
+                anyhow::anyhow!(
+                    "github is set to 'local' but GH_TOKEN not found in .capsule/.env \
+                     — add GH_TOKEN=<token> to .capsule/.env"
+                )
+            }),
+        GithubScope::Global => {
+            if let Some(token) = pre_dotenv_env.get("GH_TOKEN").filter(|t| !t.is_empty()) {
+                return Ok(token.clone());
+            }
+            let output = std::process::Command::new("gh")
+                .args(["auth", "token"])
+                .output();
+            if let Ok(out) = output {
+                if out.status.success() {
+                    let token = String::from_utf8_lossy(&out.stdout).trim().to_string();
+                    if !token.is_empty() {
+                        return Ok(token);
+                    }
+                }
+            }
+            bail!(
+                "github is set to 'global' but GH_TOKEN not found in process environment \
+                 — in CI, ensure GH_TOKEN is set by your platform \
+                 — locally, consider using 'local' instead: add GH_TOKEN to .capsule/.env"
+            )
+        }
+    }
+}
+
+// ── hooks ─────────────────────────────────────────────────────────────────────
+
+fn run_before_all(capsule_dir: &Path, env_pairs: &[(String, String)]) -> Result<()> {
+    let script = capsule_dir.join("before-all.sh");
+    if !script.exists() {
+        return Ok(());
+    }
+
+    let mut cmd = Command::new("bash");
+    cmd.arg(&script);
+    for (k, v) in env_pairs {
+        cmd.env(k, v);
+    }
+
+    let status = cmd
+        .status()
+        .with_context(|| format!("failed to run before-all.sh at {}", script.display()))?;
+
+    if !status.success() {
+        bail!(
+            "before-all.sh exited with code {}",
+            status.code().unwrap_or(-1)
+        );
+    }
+
+    Ok(())
+}
+
+// ── git ───────────────────────────────────────────────────────────────────────
+
+fn resolve_git_identity(identity: &GitIdentity, env: &HashMap<String, String>) -> (String, String) {
+    match identity {
+        GitIdentity::Capsule => ("Capsule".to_string(), "capsule@localhost".to_string()),
+        GitIdentity::User => {
+            let name = git_config_get("user.name", env);
+            let email = git_config_get("user.email", env);
+            (name, email)
+        }
+    }
+}
+
+fn git_config_get(key: &str, env: &HashMap<String, String>) -> String {
+    Command::new("git")
+        .args(["config", key])
+        .env_clear()
+        .envs(env)
+        .output()
+        .ok()
+        .filter(|o| o.status.success())
+        .and_then(|o| String::from_utf8(o.stdout).ok())
+        .map(|s| s.trim().to_string())
+        .unwrap_or_default()
+}
+
+// ── preflight ─────────────────────────────────────────────────────────────────
+
+fn check_docker() -> Result<()> {
+    let status = Command::new("docker")
+        .arg("info")
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()
+        .context(
+            "Docker is not installed or not in PATH — ensure Docker is installed and running",
+        )?;
+
+    if !status.success() {
+        anyhow::bail!(
+            "Docker daemon is not reachable — ensure Docker is running (`docker info` failed)"
+        );
+    }
+    Ok(())
+}
+
+fn env_gitignore_warning(capsule_dir: &Path) -> Option<String> {
+    let env_path = capsule_dir.join(".env");
+    if !env_path.exists() {
+        return None;
+    }
+
+    // Canonicalize env_path so git resolves it correctly regardless of current_dir.
+    // Without this, if capsule_dir is relative (e.g. ".capsule"), git would interpret
+    // ".capsule/.env" relative to its new CWD and look for ".capsule/.capsule/.env".
+    let abs_env_path = env_path.canonicalize().unwrap_or_else(|_| env_path.clone());
+    let result = Command::new("git")
+        .args(["check-ignore", "-q"])
+        .arg(&abs_env_path)
+        .current_dir(capsule_dir)
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status();
+
+    match result {
+        Ok(s) if s.success() => None,
+        _ => Some(format!(
+            "warning: {} is not gitignored — add it to .gitignore to avoid committing secrets",
+            env_path.display()
+        )),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1123,5 +1306,385 @@ mod tests {
         assert_eq!(env.len(), 2);
         assert_eq!(env["PARENT"], "79");
         assert_eq!(env["MODE"], "dry");
+    }
+
+    // ── env tests ─────────────────────────────────────────────────────────────
+
+    mod env_tests {
+        use super::super::{load_dotenv, parse_dotenv, resolve_gh_token, strip_quotes};
+        use capsule::config::GithubScope;
+        use std::collections::HashMap;
+        use tempfile::TempDir;
+
+        fn make_capsule_dir(env_content: Option<&str>) -> TempDir {
+            let dir = tempfile::tempdir().unwrap();
+            if let Some(content) = env_content {
+                std::fs::write(dir.path().join(".env"), content).unwrap();
+            }
+            dir
+        }
+
+        #[test]
+        fn parse_dotenv_basic_key_value() {
+            let env = parse_dotenv("FOO=bar\nBAZ=qux\n");
+            assert_eq!(env.get("FOO").map(|s| s.as_str()), Some("bar"));
+            assert_eq!(env.get("BAZ").map(|s| s.as_str()), Some("qux"));
+        }
+
+        #[test]
+        fn parse_dotenv_empty_content_is_empty_map() {
+            let env = parse_dotenv("");
+            assert!(env.is_empty());
+        }
+
+        #[test]
+        fn parse_dotenv_ignores_comments_and_blank_lines() {
+            let content = "# this is a comment\n\nFOO=hello\n\n# another comment\nBAR=world\n";
+            let env = parse_dotenv(content);
+            assert_eq!(env.len(), 2);
+            assert_eq!(env.get("FOO").map(|s| s.as_str()), Some("hello"));
+            assert_eq!(env.get("BAR").map(|s| s.as_str()), Some("world"));
+        }
+
+        #[test]
+        fn parse_dotenv_strips_double_quotes() {
+            let env = parse_dotenv("SECRET=\"my secret value\"\n");
+            assert_eq!(
+                env.get("SECRET").map(|s| s.as_str()),
+                Some("my secret value")
+            );
+        }
+
+        #[test]
+        fn parse_dotenv_strips_single_quotes() {
+            let env = parse_dotenv("TOKEN='abc123'\n");
+            assert_eq!(env.get("TOKEN").map(|s| s.as_str()), Some("abc123"));
+        }
+
+        #[test]
+        fn parse_dotenv_value_with_equals() {
+            let env = parse_dotenv("URL=https://example.com/path?a=1&b=2\n");
+            assert_eq!(
+                env.get("URL").map(|s| s.as_str()),
+                Some("https://example.com/path?a=1&b=2")
+            );
+        }
+
+        #[test]
+        fn strip_quotes_no_quotes_unchanged() {
+            assert_eq!(strip_quotes("hello"), "hello");
+        }
+
+        #[test]
+        fn strip_quotes_double_quotes_stripped() {
+            assert_eq!(strip_quotes("\"hello\""), "hello");
+        }
+
+        #[test]
+        fn strip_quotes_single_quotes_stripped() {
+            assert_eq!(strip_quotes("'hello'"), "hello");
+        }
+
+        #[test]
+        fn load_dotenv_absent_file_is_ok() {
+            let dir = make_capsule_dir(None);
+            assert!(load_dotenv(dir.path()).is_ok());
+        }
+
+        #[test]
+        fn resolve_gh_token_local_reads_from_dotenv_map() {
+            let pre_env: HashMap<String, String> = HashMap::new();
+            let mut dotenv: HashMap<String, String> = HashMap::new();
+            dotenv.insert("GH_TOKEN".to_string(), "ghs_localtoken".to_string());
+            let token = resolve_gh_token(&GithubScope::Local, &pre_env, &dotenv).unwrap();
+            assert_eq!(token, "ghs_localtoken");
+        }
+
+        #[test]
+        fn resolve_gh_token_local_ignores_process_env() {
+            let mut pre_env: HashMap<String, String> = HashMap::new();
+            pre_env.insert("GH_TOKEN".to_string(), "ghs_processtoken".to_string());
+            let mut dotenv: HashMap<String, String> = HashMap::new();
+            dotenv.insert("GH_TOKEN".to_string(), "ghs_dotenvtoken".to_string());
+            let token = resolve_gh_token(&GithubScope::Local, &pre_env, &dotenv).unwrap();
+            assert_eq!(token, "ghs_dotenvtoken");
+        }
+
+        #[test]
+        fn resolve_gh_token_local_missing_returns_error() {
+            let pre_env: HashMap<String, String> = HashMap::new();
+            let dotenv: HashMap<String, String> = HashMap::new();
+            let result = resolve_gh_token(&GithubScope::Local, &pre_env, &dotenv);
+            assert!(result.is_err());
+            let msg = format!("{}", result.unwrap_err());
+            assert!(msg.contains("local"), "error should mention 'local': {msg}");
+            assert!(
+                msg.contains(".capsule/.env"),
+                "error should name the file: {msg}"
+            );
+        }
+
+        #[test]
+        fn resolve_gh_token_global_reads_from_pre_dotenv_env() {
+            let mut pre_env: HashMap<String, String> = HashMap::new();
+            pre_env.insert("GH_TOKEN".to_string(), "ghs_globaltoken".to_string());
+            let dotenv: HashMap<String, String> = HashMap::new();
+            let token = resolve_gh_token(&GithubScope::Global, &pre_env, &dotenv).unwrap();
+            assert_eq!(token, "ghs_globaltoken");
+        }
+
+        #[test]
+        fn resolve_gh_token_global_missing_returns_error_or_token() {
+            let pre_env: HashMap<String, String> = HashMap::new();
+            let dotenv: HashMap<String, String> = HashMap::new();
+            match resolve_gh_token(&GithubScope::Global, &pre_env, &dotenv) {
+                Ok(_token) => {}
+                Err(e) => {
+                    let msg = format!("{e}");
+                    assert!(
+                        msg.contains("global"),
+                        "error should mention 'global': {msg}"
+                    );
+                }
+            }
+        }
+    }
+
+    // ── hooks tests ───────────────────────────────────────────────────────────
+
+    mod hooks_tests {
+        use super::super::run_before_all;
+        use std::fs;
+        use std::os::unix::fs::PermissionsExt;
+
+        #[test]
+        fn before_all_absent_is_ok() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let result = run_before_all(dir.path(), &[]);
+            assert!(
+                result.is_ok(),
+                "absent before-all.sh must return Ok: {result:?}"
+            );
+        }
+
+        #[test]
+        fn before_all_success_is_ok() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let script = dir.path().join("before-all.sh");
+            fs::write(&script, "#!/bin/sh\nexit 0\n").unwrap();
+            fs::set_permissions(&script, fs::Permissions::from_mode(0o755)).unwrap();
+
+            let result = run_before_all(dir.path(), &[]);
+            assert!(
+                result.is_ok(),
+                "before-all.sh exit 0 must return Ok: {result:?}"
+            );
+        }
+
+        #[test]
+        fn before_all_failure_is_err() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let script = dir.path().join("before-all.sh");
+            fs::write(&script, "#!/bin/sh\nexit 42\n").unwrap();
+            fs::set_permissions(&script, fs::Permissions::from_mode(0o755)).unwrap();
+
+            let result = run_before_all(dir.path(), &[]);
+            assert!(result.is_err(), "before-all.sh exit 42 must return Err");
+            let msg = format!("{:?}", result.unwrap_err());
+            assert!(
+                msg.contains("42") || msg.contains("before-all"),
+                "error must mention exit code or script name, got: {msg}"
+            );
+        }
+
+        #[test]
+        fn before_all_runs_on_host() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let script = dir.path().join("before-all.sh");
+            let sentinel = dir.path().join("ran");
+            let sentinel_str = sentinel.to_string_lossy();
+            let script_body = format!("#!/bin/sh\ntouch {sentinel_str}\nexit 0\n");
+            fs::write(&script, script_body).unwrap();
+            fs::set_permissions(&script, fs::Permissions::from_mode(0o755)).unwrap();
+
+            run_before_all(dir.path(), &[]).expect("should succeed");
+            assert!(
+                sentinel.exists(),
+                "before-all.sh should have created sentinel file"
+            );
+        }
+
+        #[test]
+        fn before_all_env_pairs_injected() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let script = dir.path().join("before-all.sh");
+            let out_file = dir.path().join("env_value.txt");
+            let out_str = out_file.to_string_lossy();
+            let script_body = format!("#!/bin/sh\necho \"$MY_TEST_VAR\" > {out_str}\n");
+            fs::write(&script, script_body).unwrap();
+            fs::set_permissions(&script, fs::Permissions::from_mode(0o755)).unwrap();
+
+            let pairs = vec![("MY_TEST_VAR".to_string(), "hello_env".to_string())];
+            run_before_all(dir.path(), &pairs).expect("should succeed");
+
+            let content = fs::read_to_string(&out_file).expect("output file should exist");
+            assert!(
+                content.trim() == "hello_env",
+                "env var should be injected into hook: got {content:?}"
+            );
+        }
+    }
+
+    // ── git tests ─────────────────────────────────────────────────────────────
+
+    mod git_tests {
+        use super::super::resolve_git_identity;
+        use capsule::config::GitIdentity;
+        use std::collections::HashMap;
+
+        fn env_with_git_config(config_path: &str) -> HashMap<String, String> {
+            let mut env = HashMap::new();
+            env.insert("GIT_CONFIG_GLOBAL".to_string(), config_path.to_string());
+            env.insert("GIT_CONFIG_NOSYSTEM".to_string(), "1".to_string());
+            env
+        }
+
+        #[test]
+        fn capsule_identity_returns_fixed_name_and_email() {
+            let env = HashMap::new();
+            let (name, email) = resolve_git_identity(&GitIdentity::Capsule, &env);
+            assert_eq!(name, "Capsule");
+            assert_eq!(email, "capsule@localhost");
+        }
+
+        #[test]
+        fn user_identity_reads_from_git_config() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let config_path = dir.path().join("gitconfig");
+            std::fs::write(
+                &config_path,
+                "[user]\n\tname = Alice Dev\n\temail = alice@example.com\n",
+            )
+            .unwrap();
+
+            let env = env_with_git_config(config_path.to_str().unwrap());
+            let (name, email) = resolve_git_identity(&GitIdentity::User, &env);
+
+            assert_eq!(name, "Alice Dev");
+            assert_eq!(email, "alice@example.com");
+        }
+
+        #[test]
+        fn user_identity_returns_empty_strings_when_git_config_missing() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let nonexistent = dir.path().join("does_not_exist");
+
+            let env = env_with_git_config(nonexistent.to_str().unwrap());
+            let (name, email) = resolve_git_identity(&GitIdentity::User, &env);
+
+            assert_eq!(name, "");
+            assert_eq!(email, "");
+        }
+
+        #[test]
+        fn user_identity_returns_empty_when_user_section_absent() {
+            let dir = tempfile::tempdir().expect("temp dir");
+            let config_path = dir.path().join("gitconfig");
+            std::fs::write(&config_path, "[core]\n\tpager = \n").unwrap();
+
+            let env = env_with_git_config(config_path.to_str().unwrap());
+            let (name, email) = resolve_git_identity(&GitIdentity::User, &env);
+
+            assert_eq!(name, "");
+            assert_eq!(email, "");
+        }
+    }
+
+    // ── preflight tests ───────────────────────────────────────────────────────
+
+    mod preflight_tests {
+        use super::super::{check_docker, env_gitignore_warning};
+        use serial_test::serial;
+        use std::fs;
+        use std::path::Path;
+        use tempfile::TempDir;
+
+        fn git_init(dir: &TempDir) {
+            std::process::Command::new("git")
+                .args(["init"])
+                .current_dir(dir.path())
+                .output()
+                .expect("git init failed");
+        }
+
+        #[test]
+        fn env_absent_returns_none() {
+            let dir = TempDir::new().unwrap();
+            assert!(env_gitignore_warning(dir.path()).is_none());
+        }
+
+        #[test]
+        fn env_present_not_gitignored_returns_warning_with_path() {
+            let dir = TempDir::new().unwrap();
+            git_init(&dir);
+            fs::write(dir.path().join(".env"), "SECRET=value").unwrap();
+            let warning = env_gitignore_warning(dir.path());
+            assert!(warning.is_some(), "expected a warning but got None");
+            let msg = warning.unwrap();
+            assert!(
+                msg.contains(".env"),
+                "warning should mention the .env path; got: {msg}"
+            );
+        }
+
+        #[test]
+        fn env_present_and_gitignored_returns_none() {
+            let dir = TempDir::new().unwrap();
+            git_init(&dir);
+            fs::write(dir.path().join(".env"), "SECRET=value").unwrap();
+            fs::write(dir.path().join(".gitignore"), ".env\n").unwrap();
+            assert!(env_gitignore_warning(dir.path()).is_none());
+        }
+
+        #[test]
+        #[serial]
+        fn env_relative_capsule_dir_gitignored_returns_none() {
+            let root = TempDir::new().unwrap();
+            git_init(&root);
+            let capsule_dir = root.path().join(".capsule");
+            fs::create_dir(&capsule_dir).unwrap();
+            fs::write(capsule_dir.join(".env"), "SECRET=value").unwrap();
+            fs::write(root.path().join(".gitignore"), ".capsule/.env\n").unwrap();
+
+            let original = std::env::current_dir().unwrap();
+            std::env::set_current_dir(root.path()).unwrap();
+            let result = env_gitignore_warning(Path::new(".capsule"));
+            std::env::set_current_dir(original).unwrap();
+
+            assert!(result.is_none(), "expected no warning but got: {result:?}");
+        }
+
+        #[test]
+        #[serial]
+        fn env_relative_capsule_dir_not_gitignored_returns_warning() {
+            let root = TempDir::new().unwrap();
+            git_init(&root);
+            let capsule_dir = root.path().join(".capsule");
+            fs::create_dir(&capsule_dir).unwrap();
+            fs::write(capsule_dir.join(".env"), "SECRET=value").unwrap();
+
+            let original = std::env::current_dir().unwrap();
+            std::env::set_current_dir(root.path()).unwrap();
+            let result = env_gitignore_warning(Path::new(".capsule"));
+            std::env::set_current_dir(original).unwrap();
+
+            assert!(result.is_some(), "expected a warning but got None");
+        }
+
+        #[test]
+        #[capsule_macros::requires_docker]
+        fn docker_available_check_succeeds() {
+            check_docker().expect("docker check should succeed when Docker is running");
+        }
     }
 }
